Make project intelij compatible
