--- .metadata/.plugins/org.eclipse.m2e.logback/0.log
@@ -1,1 +1,0 @@
-2024-06-15 22:10:19,757 [Worker-4: Loading available Gradle versions] INFO  o.e.b.c.i.u.g.PublishedGradleVersions - Gradle version information cache is up-to-date. Trying to read.
