--- .metadata/version.ini
@@ -1,3 +1,0 @@
-#Sat Jun 15 22:10:01 IST 2024
-org.eclipse.core.runtime=2
-org.eclipse.platform=4.27.0.v20230302-0300
