+++ src/main/java/com/app/ayon/QuizApp/controller/QuestionController.java
@@ -1,0 +1,26 @@
+package com.app.ayon.QuizApp.controller;
+
+
+import com.app.ayon.QuizApp.model.Question;
+import com.app.ayon.QuizApp.service.QuestionService;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+import java.util.List;
+
+@RestController
+@RequestMapping("question")
+public class QuestionController {
+
+    @Autowired
+    QuestionService questionService;
+
+    @GetMapping("allQuestions")
+    public List<Question> getAllQuestions() {
+       return questionService.getAllQuestions();
+    }
+
+
+}
