--- src/main/java/com/ayon/quizapp/controller/QuestionController.java
@@ -1,26 +1,0 @@
-package com.ayon.quizapp.controller;
-
-import java.util.List;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-import com.ayon.quizapp.model.Question;
-import com.ayon.quizapp.service.QuestionService;
-
-@RestController
-@RequestMapping("question")
-public class QuestionController {
-	
-	@Autowired
-	QuestionService questionService;
-	
-	@GetMapping("allQuestions")
-	public List<Question> getAllQuestions() {
-		return questionService.getAllQuestions();
-	}
-	
-
-}
