+++ src/main/resources/application.properties
@@ -1,15 +1,10 @@
 spring.application.name=QuizApp
 
 
-spring.datasource.url=jdbc:postgresql://localhost:5432/questionDB	
+spring.datasource.url=jdbc:postgresql://localhost:5432/questionDB
 spring.datasource.username=postgres
 spring.datasource.password=1234
 spring.datasource.driver-class-name=org.postgresql.Driver
 
-# Hibernate ddl auto (create, create-drop, update): with "create-drop" the database
-# schema will be automatically created afresh for every start of application
 spring.jpa.hibernate.ddl-auto=update
-
-
-# Allows Hibernate to generate SQL optimized for a particular DBMS
 spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.PostgreSQLDialect
