+++ src/test/java/com/app/ayon/QuizApp/QuizAppApplicationTests.java
@@ -1,4 +1,4 @@
-package com.ayon.quizapp;
+package com.app.ayon.QuizApp;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
